Database Design to P2SH + Signature-based
Due to the fact that LN Tranactions are not considered standard, it is
necessary to use P2SH.

To not save all outputscripts, we further want to be able to reconstruct
the script. The database should reflect this.

Changed: We save the current signature directly in the channels table.
Changed: The channel now has a version number, to keep track of which
payments where included together with which revocation hashes.
Removed; The transaction table, it is longer necessary to save
transactions, we can build all transactions and just add the signature.