Implemented dual-tx approach
LNPaymentCMessage will now include a list with signatures for each
payment.

The channel transaction of each party does now pay to multisig for
payments that they are entitled to normally. They receive a signature
for spending from that output that pays towards a revocable
transaction.

Using this approach, we have decoupled refund time from revocation time.