Refactored ConnectionManager / ChannelManager
Previously the ConnectionManager was also in charge of creating new
channels. This turned out to be super messy, with various callbacks to
make sure to keep open connections with nodes that we have open
channels with, but not reopening connections after closing a channel.

This change moved that responsibility towards the ChannelManager, which
will now employ a dedicated worker thread to keep track of open
channels / connections and tries to reconnect if necessary.