Fix: Use timestamp to calculate the hash for P2PDataObjects
Also drastically reduce the MAXIMUM_AGE_SYNC_DATA to ten minutes, which
will effect how often data will be broadcasted and when we prune data
out of our database. This should scale to quite a few hundred peers
very easily, integrating a framework that is responsible for the gossip
protocol is a next step though, as our current custom solution is not
very efficient.