Change anchor design to 2-of-2 multis
With SegWit approaching fairly quickly it makes more sense to utilise
the malleability-free environment it brings. This commit changes from
the anchor design described in the ‘deployable lightning’ paper towards
a plain 2-of-2 multisig.

For doing so both parties mutually sign a transaction funding the
channel. They will both add inputs for funding and outputs for the
remaining change. They then create the first commitment that also
serves as a refund in case either party disappears directly after
broadcasting the anchor. After the signatures for the channel has been
exchanged, the parties exchange all necessary signatures for their
inputs.

For this to work we need SegWit, as otherwise one cannot construct and
sign the channel without exchanging signatures for the anchor.