Updated payment procedure to handle conflicts properly
When both parties send a request to update the channel concurrently,
there was a problem that could have led to loss of funds.

We now send a fresh revocation hash to be used for the next exchange at
the beginning of the exchange (with the A message), such that in case
of a conflict, we can use that unused revocation hash directly to
construct the new status.